update to define index

+++ meta_data/create_file_paths_table.sql
@@ -1,6 +1,6 @@
 \connect bathymetry
 CREATE TABLE file_paths (
-        source_id float8 NOT NULL ,
+        source_id int4 NOT NULL ,
         file_path varchar(255) NOT NULL
     );
 COMMIT;
